Topic data structure amended, in middle of parsing BBC
The topics list has been amended to a dictionary of dictionaries. The keys are the topics themselves, which map to a dictionary of sources and the links to their specific pages dedicated to that topic. Functions to build these links in page parser have been removed subsequently, everything instead using extractPageData. Currently in middle of extracting BBC article links, this is proving harder because of articles appearing in multiple topic pages and not having a consistent link structure. Will likely move summarising and translating fuctionality to separate functions later