Added Reuters and halfway through manual text extraction
Reuters has been added and is being summarised and read fine. Only issue with that is that the Reuters articles themselves contain a place of publishing and a publisher at the start of every article but this can be removed.

Halfway through the text extraction.

Ran into issue where topic name was not the same for all sources, e.g. topic "uk" for Guardian is no good, it need to be "uk-news". May fix this later by having a map for this. Right now, is just hard-coded.