Finished extracting AP text. Made breakthrough with HTML tag removal
Found a piece of code online that seems to do the job 100% of the time in removing tags from text containing HTML. This includes tags with classes and additional attributes. Tested it on other functions struggling to remove links to amazing success.
Had a downside whereby Reuters is now returning additional text at the end it shouldn't. Also noticed now that old news is being fetched from some pages but that's just cos the page is showing it. For example, got an article from December 13. Might be worth reading out the date of the article in TTS