Fixed issue with some topics not being covered. Also fixed AP issues
There was a long-standing issue in bulletin where, if an article could not be found for a topic, the do-while loop wouldn't work, likely because of async code. Fixed this so that the error is caught and then a recursive function is called to retry until a success or a timeout of 10 tries.

Also found a bug where SMMRY wasn't extracting a headline from an AP article. This is because AP does not make use of the <title> tag. Made a function in summarise.mjs specifically to extract an AP headline.