Got a dynamic form, only missing correct values being set on clicking icon
Currently, whenever the icon is set, the page is in its default state, i.e. if a bulletin is playing then the play button is shown as not playing when we need it to show the pause option.

Still making use of messages in popup.mjs but can probably change this to straight up replacing uifunction.mjs functionality soon.

Also modified the way bulletin.mjs works in order to more dynamically interact with elements in it (hopefully). Articles, when fetched, are added to a list. Once all articles are fetched, the list is read in order. Ideally, if a change is received during playback, the remaining articles can reflect this change.