WEIRD BUG APPEARS IN THIS COMMIT
Caps to emphasise.

Wish to make clear that this code WORKS. However - some difference hidden away somewhere in the code, most likely popup.mjs, means that errors are flagged in the console but they don't actually occur.

Errors are to do with being unable to find html elements. This is despite checking the DOM is FULLY loaded. Completely unable to understand why. This is code that ran perfectly well in the previous commit and that I haven't touched.

The errors are flagged when the extension is first loaded. From that moment on, they mostly disappear but obviously remain in the console, which is very annoying.  Except with the stop button. I have no idea.

Because it doesn't break functionality and it's beyond my energy levels, I'm giving up for now.