Slightly improved UI, removed AP headline bug, broken headline up with text splitter
Headlines could potentially also be too long. To prevent this, they are broken up the same way as text now. If language is English, the headline and text are read out with no breaks.

AP headline bug also fixed: for some reason SMMRY can't extract it so we use our manual extractor

UI is slightly improved so that controls and heading are centred